Fixed all unecessary commits
Commits currently are dreadful
making it hard for us to see what changed
and why do we do that change.

We want to go back to start to create a story of
commits but this is the best we could.

We just have to start all over again on comminting messages.

The last story feature created

Resolve: Story#51